+++ cryptography/hedera-cryptography-hinTS/src/test/java/com/hedera/cryptography/hints/HintsLibraryBridgeTest.java
@@ -38,7 +38,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testComputeHints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         final byte[] secretKey = INSTANCE.generateSecretKey(HintsConstants.RANDOM_2);
@@ -51,7 +51,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testComputeHintsConstraints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         final byte[] secretKey = INSTANCE.generateSecretKey(HintsConstants.RANDOM_2);
 
         // n must be power of two
@@ -97,7 +97,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testValidateHintsKey() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         final byte[] secretKey = INSTANCE.generateSecretKey(HintsConstants.RANDOM_2);
@@ -111,7 +111,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testValidateHintsConstraints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         final byte[] secretKey = INSTANCE.generateSecretKey(HintsConstants.RANDOM_2);
         final byte[] hints = INSTANCE.computeHints(crs, secretKey, 2, 4);
 
@@ -157,7 +157,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testPreprocess() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         final byte[] secretKey = INSTANCE.generateSecretKey(HintsConstants.RANDOM_2);
@@ -174,7 +174,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testPreprocessConstraints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
 
         final byte[] secretKey0 = INSTANCE.generateSecretKey(HintsConstants.RANDOM_0);
         final byte[] hints0 = INSTANCE.computeHints(crs, secretKey0, 0, 4);
@@ -274,7 +274,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testVerifyBls() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         final byte[] secretKey = INSTANCE.generateSecretKey(HintsConstants.RANDOM_2);
@@ -297,7 +297,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testVerifyBlsConstraints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         final byte[] secretKey = INSTANCE.generateSecretKey(HintsConstants.RANDOM_2);
         final int partyId = 2;
         final byte[] hints = INSTANCE.computeHints(crs, secretKey, partyId, 4);
@@ -335,7 +335,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testVerifyBlsBatch() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         final int partyId0 = 0;
@@ -361,7 +361,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testVerifyBlsBatchConstraints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
 
         // partyId 0
         final byte[] secretKey0 = INSTANCE.generateSecretKey(HintsConstants.RANDOM_0);
@@ -417,7 +417,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testAggregateSignatures() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         // partyId 2
@@ -447,7 +447,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testAggregateSignaturesConstraints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
 
         // partyId 0
         final byte[] secretKey0 = INSTANCE.generateSecretKey(HintsConstants.RANDOM_0);
@@ -528,7 +528,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testVerifyAggregate_meetsThreshold() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         // partyId 2
@@ -564,7 +564,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testVerifyAggregate_doesNotMeetThreshold() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
         assertArrayEquals(CRSConstants.INIT_CRS_4, crs);
 
         // partyId 2
@@ -600,7 +600,7 @@ public class HintsLibraryBridgeTest {
 
     @Test
     void testVerifyAggregateConstraints() {
-        final byte[] crs = INSTANCE.initCRS(4);
+        final byte[] crs = INSTANCE.initCRS((short) 4);
 
         final byte[] secretKey0 = INSTANCE.generateSecretKey(HintsConstants.RANDOM_0);
         final byte[] hints0 = INSTANCE.computeHints(crs, secretKey0, 0, 4);
